Unify the player surface; fix CompareView's drift
The AVPlayerLayer host was copy-pasted into three views (Practice,
Compare, Trim) and only Practice's copy had the background
detach/reattach logic — so compare/trim playback died when the app
backgrounded. CompareView had also rotted into a stale fork of
PracticeView. Code review surfaced the cluster; this collapses it.

- New Views/PlayerSurface.swift: one shared surface with the
  background handling, plus a fix to the observer registration
  (clear before re-adding so didMoveToWindow firing twice can't stack
  duplicate NotificationCenter observers).
- Deleted the three private clones (PlayerSurface/PlayerView in
  PracticeView, ComparePlayerSurface/CompareLayerView in CompareView,
  TrimPlayerSurface/TrimPlayerView in TrimView); all three now use the
  shared surface.
- CompareView: load preferredLocalFileURL (trim > original) instead of
  trimmedFileURL — untrimmed clips were falling back to a PHAsset
  round-trip, and clips whose Photos original was deleted failed to
  load entirely (worked in Practice, not Compare).
- CompareView: replace .aspectRatio(16/9) with a fill so AVPlayerLayer
  letterboxes once instead of double-letterboxing non-16:9 clips
  (same fix as PracticeView #67).

168 tests pass.
