+++ StepBack/Views/PracticeView.swift
@@ -638,77 +638,6 @@ extension PracticeView {
     }
 }
 
-// MARK: - Player surface
-
-private struct PlayerSurface: UIViewRepresentable {
-    let player: AVPlayer
-
-    func makeUIView(context: Context) -> PlayerView {
-        let view = PlayerView()
-        view.playerLayer.player = player
-        view.playerLayer.videoGravity = .resizeAspect
-        return view
-    }
-
-    func updateUIView(_ uiView: PlayerView, context: Context) {
-        uiView.playerLayer.player = player
-    }
-}
-
-private final class PlayerView: UIView {
-    override static var layerClass: AnyClass { AVPlayerLayer.self }
-
-    var playerLayer: AVPlayerLayer {
-        guard let layer = layer as? AVPlayerLayer else {
-            preconditionFailure("PlayerView.layer must be an AVPlayerLayer")
-        }
-        return layer
-    }
-
-    // Detach the AVPlayer from the layer when backgrounding so iOS keeps
-    // audio flowing even with the Background Audio capability enabled.
-    // While a video output is attached, the system aggressively pauses the
-    // player; nil-ing it lets the audio session keep going. Reattach on
-    // foreground so the user sees frames again.
-    private var stashedPlayer: AVPlayer?
-
-    override func didMoveToWindow() {
-        super.didMoveToWindow()
-        if window != nil {
-            NotificationCenter.default.addObserver(
-                self,
-                selector: #selector(detachForBackground),
-                name: UIApplication.didEnterBackgroundNotification,
-                object: nil
-            )
-            NotificationCenter.default.addObserver(
-                self,
-                selector: #selector(reattachForForeground),
-                name: UIApplication.willEnterForegroundNotification,
-                object: nil
-            )
-        } else {
-            NotificationCenter.default.removeObserver(self)
-        }
-    }
-
-    @objc private func detachForBackground() {
-        stashedPlayer = playerLayer.player
-        playerLayer.player = nil
-    }
-
-    @objc private func reattachForForeground() {
-        if let stashedPlayer {
-            playerLayer.player = stashedPlayer
-        }
-        stashedPlayer = nil
-    }
-
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-    }
-}
-
 // MARK: - Scrubber
 
 private struct Scrubber: View {
